Export threaded sending classes
