+++ src/settings.json
@@ -6,7 +6,7 @@
     "attach": "",
     "multithread": null,
     "title": "SpamBotFromHell",
-    "debug": false,
+    "debug": true,
     "delay": 0,
     "max_retries": 5,
     "display_from": "spammer@spam.com",
